requirements.txt
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.24
torch>=2.0
python-dotenv>=1.0
